[-Wsign-compare]
/p/vast1/rountree/repos/Spindle/src/client/biter/../../biter/shmutil.c:45:43:
Split out pid_t and locks into their own variables/arrays.
Affects lock_t, base_header_t, biter_header_t, and shmcache_header_t.

/p/vast1/rountree/repos/Spindle/src/client/biter/../../biter/demultiplex.c:197:17:
Changed type of msg_header_t.msg_target from uint32_t to int (src/biter/demultiplex.h)

/p/vast1/rountree/repos/Spindle/src/client/auditclient/writablegot.c:121:21:
Cribbed from earlier solution.

/p/vast1/rountree/repos/Spindle/src/client/client/../../utils/parseloc.c:250:15:
Cribbed.

/p/vast1/rountree/repos/Spindle/src/client/client/../../utils/fileutil.c:84:25:
Cribbed.

/p/vast1/rountree/repos/Spindle/src/client/client/remap_exec.c:94:33:
Cribbed.

/p/vast1/rountree/repos/Spindle/src/client/client/intercept_readlink.c:147:12:
Cribbed.

/p/vast1/rountree/repos/Spindle/src/client/client_comlib/client_api_socket.c:237:11:
Cribbed early solution.

/p/vast1/rountree/repos/Spindle/src/client/client_comlib/client_api.c:235:22:
Trickier than most, as the code is dependent on the size of the
variables in question.

/p/vast1/rountree/repos/Spindle/src/client/client/realpath.c:369:17:
In the interest of tracking the original glibc code, disable this warning.

/p/vast1/rountree/repos/Spindle/src/client/subaudit/update_pltbind.c:185:18:
Cribbed.

/p/vast1/rountree/repos/Spindle/src/server/biter/../../biter/biterd.c:90:18:
Cribbed.

/p/vast1/rountree/repos/Spindle/src/server/cache/global_name.c:71:11:
int->size_t

/p/vast1/rountree/repos/Spindle/src/server/cobo/../../cobo/cobo.c:477:25:
Replaced inet_addr() call with intel_aton() per man page recommendation.

/p/vast1/rountree/repos/Spindle/src/server/cobo/../../cobo/handshake.c:689:22: warning:
Used ssize_t to allow errors and data to share the same channel.

/p/vast1/rountree/repos/Spindle/src/server/comlib/ldcs_api_socket.c:291:11:
Duplicated code...

/p/vast1/rountree/repos/Spindle/src/server/comlib/ldcs_api_pipe.c: In function ‘ldcs_send_msg_pipe’:
Familiar issue of read/write returning ints and those sizes being stored in size_ts.

/p/vast1/rountree/repos/Spindle/src/server/auditserver/ldcs_elf_read.c: In function ‘readUpTo’:
Interesting that fread does not return an error condition, nor does it set errno.o

/p/vast1/rountree/repos/Spindle/src/server/auditserver/ldcs_audit_server_requestors.c: In function ‘get_requestor’:
Cribbed.

/p/vast1/rountree/repos/Spindle/src/server/auditserver/ldcs_audit_server_numa.c: In function ‘pattern_match’:
Cribbed.

src/server/auditserver/ldcs_audit_server_handlers.c
Nothing remarkable.

Passes runTests.

+++ src/include/ldcs_api_pipe.h
@@ -35,8 +35,8 @@ char *ldcs_get_connection_string_pipe(int fd);
 int ldcs_register_connection_pipe(char *connection_str);
 
 /* internal */
-int _ldcs_write_pipe(int fd, const void *data, int bytes );
-int _ldcs_read_pipe(int fd, void *data, int bytes, ldcs_read_block_t block );
+size_t _ldcs_write_pipe(int fd, const void *data, int bytes );
+size_t _ldcs_read_pipe(int fd, void *data, int bytes, ldcs_read_block_t block );
 
 typedef enum {
    LDCS_PIPE_FD_TYPE_SERVER,
